Proposal: Auto-run recordings if Race Menu never opened
Should a user use the `CenterOnCell` or `CenterOnWorld` commands to leave the main menu, recordings set to automatically run will not, in these cases, be run. I am not sure if this is the desired behavior. Perhaps additional logic could be added to determine if the recording should run on COC/COW?

Note: I have not tested this, nor have I compiled the script.